Refactor DateTimeTools and introduce AiToolsService interface for better tool management
