+++ src/main/java/space/parzival/discord/badley/ai/generic/DateTimeTools.java
@@ -1,15 +1,16 @@
-package space.parzival.discord.badley.ai;
+package space.parzival.discord.badley.ai.generic;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.ai.tool.annotation.Tool;
 import org.springframework.stereotype.Component;
+import space.parzival.discord.badley.ai.AiToolsService;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 
 @Slf4j
 @Component
-public class DateTimeTools {
+public class DateTimeTools implements AiToolsService {
     @Tool(description = "Get the current date and time in ISO datetime format.")
     public String getCurrentDateTime() {
         log.info("Get the current date and time in ISO datetime format.");
