+++ src/main/java/space/parzival/discord/badley/configuration/SpringAIConfiguration.java
@@ -1,14 +1,18 @@
 package space.parzival.discord.badley.configuration;
 
 import lombok.AllArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.ai.chat.client.ChatClient;
 import org.springframework.ai.chat.client.advisor.MessageChatMemoryAdvisor;
 import org.springframework.ai.chat.memory.ChatMemory;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import space.parzival.discord.badley.ai.DateTimeTools;
+import space.parzival.discord.badley.ai.AiToolsService;
 import space.parzival.discord.badley.configuration.properties.AiProperties;
 
+import java.util.List;
+
+@Slf4j
 @Configuration
 @AllArgsConstructor
 public class SpringAIConfiguration {
@@ -17,14 +21,15 @@ public class SpringAIConfiguration {
     private AiProperties aiProperties;
 
     // ai tools
-    private final DateTimeTools aiDateTimeTools;
+    private final List<? extends AiToolsService> aiToolsServices;
 
     @Bean
     public ChatClient chatClient() {
+        log.info("Registering {} AI tools", aiToolsServices.size());
         return chatClientBuilder
                 .defaultAdvisors(new MessageChatMemoryAdvisor(chatMemory))
                 .defaultSystem(aiProperties.getPersonality())
-                .defaultTools(aiDateTimeTools)
+                .defaultTools(aiToolsServices.toArray())
                 .build();
     }
 }
